requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=1.0.0
schedule>=1.2.0
pytz>=2024.1
feedparser>=6.0.0
